refactor(chat): 자료형 변경 (Long -> Integer)

+++ src/main/java/org/com/dungeontalk/domain/chat/dto/PresenceBroadcastDto.java
@@ -9,14 +9,14 @@ import lombok.Getter;
 @Builder
 public class PresenceBroadcastDto {
     private String roomId;
-    private Long maxCapacity;
+    private Integer maxCapacity;
     private long connectedCount;
     private List<PresenceMessageDto.MemberPresence> members; // 기존 타입 재활용
     private String eventType;           // JOIN, LEAVE, JOIN_IGNORED, LEAVE_IGNORED
     private Instant eventTime;
 
     public static PresenceBroadcastDto of(
-        String roomId, Long maxCapacity, long connectedCount,
+        String roomId, Integer maxCapacity, long connectedCount,
         List<PresenceMessageDto.MemberPresence> members, String eventType
     ) {
         return PresenceBroadcastDto.builder()
